gcp_benchmarks/dna_visualization/requirements.txt
dna-features-viewer
google-cloud-storage
matplotlib
pyyaml==6.0.1
pytz==2024.1
zstandard==0.23.0
